requirements.txt
spotipy
spotdl
pyyaml
colorama
